+++ examples/scripts/meshing/partitioning_perp_plates_and_beams.py
@@ -22,10 +22,13 @@ def main():
     pl3 = ada.Plate("pl3", p1x1, 0.01, orientation=ada.Placement(xdir=(1, 0, 0), zdir=(0, -1, 0)))
 
     p = ada.Part("MyFem") / [pl1_5, pl3, bm_1, bm_2]
+
     fem = p.to_fem_obj(1, bm_repr="line", pl_repr="shell")
+    fem.show(solid_beams=True)
     p.fem = fem
+
     a = ada.Assembly("Test") / p
-    a.to_fem("ADA_pl_mesh", "usfos", overwrite=True)
+    # a.to_fem("ADA_pl_mesh", "usfos", overwrite=True)
 
     assert len(fem.nodes) == 10
     assert len(fem.elements) == 13
